refactor: move AppBar to features folder
